backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.0.0
fastapi==0.104.1
h11==0.14.0
idna==3.10
motor==3.3.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2
Werkzeug
pypdf
anthropic
requests
beautifulsoup4
langchain
langchain-anthropic
selenium
webdriver-manager